Credit Card Validator v1.2.0 CreditCardValidator.CLI

Parse command line args, dispatch to worker functions to perform validation.

Functions

format_output(arg)

Specs

format_output(CreditCardValidator.card_validation) :: String.t

Formats a line of output after card has been validated.

Example

iex> CreditCardValidator.CLI.format_output({:VISA, "4111111111111111", :valid})
"VISA: 4111111111111111       (valid)"

help()

Specs

help :: :ok

Shows the help screen.

main(argv)

Specs

main([String.t]) :: :ok

escript application entry point. When argv is an empty list, lines from stdin will be processed. Otherwise if -h, —help or any other unrecognised flags are given help will be shown.

process()

Specs

process :: :ok

Stream lines containing credit cards from stdin, validates each one and outputs a formatted table to stdout.