Credit Card Validator v1.2.0 CreditCardValidator

Defines the core domain logic for validating credit cards.

A credit card is given simply as a String, which may contain spaces.

validate_card/1 can be used to find the type and validity of a card. card_type/1 can be used to only find the type of a card. luhn/1 can be used to only validate the digits.

See CreditCardValidator.CLI module for processing credit cards in batch from stdio.

Note that the card_type logic is baked into the code of this module, not externalized into a config. This should be fine, as erlang/elixir supports module updating in running applications. As new types are added, simply create additional clauses for card_type/1 and reload the module.

Types

card_number()

card_number() :: String.t

card_type()

card_type() :: :AMEX | :Discover | :MasterCard | :VISA | :Unknown

card_validation()

card_validation() :: {card_type, card_number, validation}

validation()

validation() :: :valid | :invalid

Functions

card_type(arg1)

card_type([integer]) :: card_type

Pattern match on the leading digits and length of a credit card number to determining the type.

:AMEX -> 15 digits beginning with 3,4 or 3,7

:Discover -> 16 digits beginning with 6,0,1,1

:MasterCard -> 16 digits beginning with 5,1-5

:Visa -> 13 or 16 digits beginning with 4

:Unknown -> otherwise

Example

iex> CreditCardValidator.card_type([4,2,3,4,5,6,7,8,9,0,1,2,3])
:VISA

iex> CreditCardValidator.card_type([1,2,3,4,5])
:Unknown

digit_list(card)

digit_list(card_number) :: [integer]

Convert a credit card number string into a list of integer digits. This is later used for pattern matching and luhn validation.

Example

iex> CreditCardValidator.digit_list("3754 3210 987 654 321")
[3, 7, 5, 4, 3, 2, 1, 0, 9, 8, 7, 6, 5, 4, 3, 2, 1]

digit_string(card)

digit_string(card_number) :: card_number

Removes all whitespace from a credit card number, producing a large integer string.

Example

iex> CreditCardValidator.digit_string("3754 3210 987 654 321")
"37543210987654321"

luhn(digits)

luhn([integer]) :: validation

Validate a credit card number using the Luhn algorithm. The input must already be split into a digit list, eg with Integer.digits or CreditCardValidator.digit_list See https://en.wikipedia.org/wiki/Luhn_algorithm for a complete description.

Example

iex> CreditCardValidator.luhn([4,4,0,8,0,4,1,2,3,4,5,6,7,8,9,3])
:valid

iex> CreditCardValidator.luhn([4,4,0,8,0,4,1,2,3,4,5,6,7,8,9,4])
:invalid

validate_card(card)

validate_card(card_number) :: card_validation

Given a string representing a credit card number, returns a tuple containing card type atom, card number string, and atom for validity of the card. Card type rules are defined by the card_type/1 function. Validity rules are defined by the luhn/1 function.

Example

iex> CreditCardValidator.validate_card("3434 5678 9012 345")
{:AMEX, "3434 5678 9012 345", :invalid}